🚚 rename `visitor` to `adapter`
