+++ examples/visual_math/src/app/graph/adapter.rs
